Only open modal once recent logs are available
